refactor(config): dedupe observe suspicion checks
