+++ src/config/io.observe-suspicious.ts
@@ -1,0 +1,50 @@
+import { isRecord } from "../utils.js";
+
+export type ConfigObserveSuspiciousBaseline = {
+  bytes: number;
+  hasMeta: boolean;
+  gatewayMode: string | null;
+};
+
+function isUpdateChannelOnlyRoot(value: unknown): boolean {
+  if (!isRecord(value)) {
+    return false;
+  }
+  const keys = Object.keys(value);
+  if (keys.length !== 1 || keys[0] !== "update") {
+    return false;
+  }
+  const update = value.update;
+  if (!isRecord(update)) {
+    return false;
+  }
+  const updateKeys = Object.keys(update);
+  return updateKeys.length === 1 && typeof update.channel === "string";
+}
+
+export function resolveConfigObserveSuspiciousReasons(params: {
+  bytes: number;
+  hasMeta: boolean;
+  gatewayMode: string | null;
+  parsed: unknown;
+  lastKnownGood?: ConfigObserveSuspiciousBaseline;
+}): string[] {
+  const reasons: string[] = [];
+  const baseline = params.lastKnownGood;
+  if (!baseline) {
+    return reasons;
+  }
+  if (baseline.bytes >= 512 && params.bytes < Math.floor(baseline.bytes * 0.5)) {
+    reasons.push(`size-drop-vs-last-good:${baseline.bytes}->${params.bytes}`);
+  }
+  if (baseline.hasMeta && !params.hasMeta) {
+    reasons.push("missing-meta-vs-last-good");
+  }
+  if (baseline.gatewayMode && !params.gatewayMode) {
+    reasons.push("gateway-mode-missing-vs-last-good");
+  }
+  if (baseline.gatewayMode && isUpdateChannelOnlyRoot(params.parsed)) {
+    reasons.push("update-channel-only-root");
+  }
+  return reasons;
+}
